ksmbd-tools: fix potential null dereference of share->comment
share->comment can be null, depending on the config file. This is
already handled in line 61 in the same file but missing here.

Skip any operation on share->comment if it is null. This fixes a problem
when clients cannot connect to the ksmbd server anymore as ksmbd.mountd
fails with the following error in the kernel log:
  ksmbd: handle_startup_event:381: Reconnect to a new user space daemon
  do_page_fault(): sending SIGSEGV to pool for invalid read access from 00000000
  epc = 77ebaff4 in libc.so[77e42000+9c000]
  ra  = 0040dc3b in ksmbd.mountd[400000+15000]
  ksmbd: handle_startup_event:381: Reconnect to a new user space daemon

Fixes: 8ec5ca70e8bb87 ("ksmbd-tools: add max_len and actual_len arguments in ndr_write_vstring()")
Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>