ksmbd-tools: fix meson build
The HAVE_KRB5_AUTH_CON_GETRECVSUBKEY macro block deals with the
krb5_auth_con_getrecvsubkey symbol not being exported even though it is
defined in Heimdal. It is *not* sufficient to check whether it has a
declaration in the krb5.h header file.

The HAVE_KRB5_AUTH_CON_GETAUTHENTICATOR_DOUBLE_POINTER macro block
deals with the type difference in the krb5_authenticator parameter to
krb5_auth_con_getauthenticator. For an unknown reason, this check was
hardcoded as true in the meson build.

Signed-off-by: atheik <atteh.mailbox@gmail.com>