ksmbd-tools: fix printing of debug component status
The debug option of ksmbd.control toggles a debug component by doing a
write() to /sys/class/ksmbd-control/debug. It then does a read() to get
status of all components but does not seek back to the beginning. This
leads to an incomplete status line. Here, status of the smb component
is only partially printed:

# ksmbd.control -d all
[ksmbd.control/134655]: INFO: b] [auth] [vfs] [oplock] [ipc] [conn] [rdma]

Seek back to the beginning after write() with lseek().

Signed-off-by: atheik <atteh.mailbox@gmail.com>