ksmbd-tools: improve usage strings and argument handling
The usage strings were terse and outdated. They also followed an
inconsistent and unusual format. Rewrite the usage strings to follow a
more common convention found in GNU programs and others. Include
compile-time constants in the usage strings as the programs themselves
do not have manual pages.

As usage() is only ever called from main(), the use of exit() is not
necessary. Instead, let main() return after calling usage(). It is
also unusual that printing the usage page would denote a program error.
Instead, only return with EXIT_FAILURE after usage() when the user
provides an invalid option or is missing an argument.

As show_version() is only ever called from main(), the use of exit()
is not necessary. Instead, let main() return after calling
show_version(). It is unusual that showing program version would denote
a program error. Return EXIT_SUCCESS after show_version().

It is unusual to print the complete usage page if the user provides an
invalid option or an argument is missing. Instead, show the user that
the usage page can be accessed through the help option.

When ksmbd.adduser was invoked without arguments, it hit the user name
sanity check error, which is confusing. Instead, show the user how to
access the usage page. When given either the password or import-users
option alone, it also hit the error. Instead, give a descriptive error
message. Also, document what user names and passwords are valid, that
the verbose option is not implemented, and that update-user and
del-user require restarting ksmbd for changes to take effect. Note that
import-users has a misleading name, which implies that it does
something by itself.

When ksmbd.addshare was invoked without arguments, it hit the missing
options option error, which is confusing, and then printed the usage
page. There was also a return statement after usage() in main(), which
was never reached. Instead, show the user how to access the usage page.
When given either the options or config option alone, it also hit the
error. Instead, give a descriptive error message. The usage page gave
no hint that the add-share and update-share options both require the
options option. When hitting this condition, show a descriptive error
message. Document this, what share names are valid, that the verbose
option is not implemented, and that update-share and del-share require
restarting ksmbd for changes to take effect.

When ksmbd.control was invoked without arguments, it printed the usage
page. Instead, show the user how to access the usage page. Document
that enabled debug components are in brackets.

For ksmbd.mountd, return from main() instead of calling exit(). Also,
document that nodetach has an optional argument which determines
whether to become a process group leader.

For all programs, show the user how to access the usage page if there
are non-option arguments.

Signed-off-by: atheik <atteh.mailbox@gmail.com>