ksmbd-tools: fix password prompting for ksmbd.adduser
Closing stdin with ^D when being prompted for a password resulted in
nonsensical error messages. Instead, detect it and ask for a password
again.

Password comparison is done with memcmp() over the entire buffers
instead of using strcmp() to compare up to the null bytes. It was then
impossible to give a shorter password after an attempt to give a longer
password, since part of the longer password was still in the buffer.
Instead of switching to strcmp(), zero the buffers at the beginning of
the prompting loop as that is already necessitated by the first fix in
this commit.

While we're at it, fix the check for an empty password, which seemed to
erroneously assume that it still contained a newline.

Signed-off-by: atheik <atteh.mailbox@gmail.com>