ksmbd-tools: rewrite readme
The plain text README and its Markdown version README.md were
inconsistent and outdated. There was a reference to users.db instead
of ksmbdpwd.db and the example share options argument created a
malformed smb.conf. Since the usage strings are now more complete as
of commit 47d93fc, repurpose the README files to give an example usage
session which setups a share and a user, briefly introduces smb.conf,
and shows an example about the current limitation of updating and
deleting users and shares.

Signed-off-by: atheik <atteh.mailbox@gmail.com>