ksmbd-tools: create ksmbd sysconfdir in the install target
The install target did not create the ksmbd directory under sysconfdir,
leaving it up to the user. Modify the autotools and meson builds such
that it is created in the install target. So as to not leave it empty,
install smb.conf.example there. Also, add contents of configuration.txt
as a comment on top of smb.conf.example. Rename the example homes share
to tmp.

Signed-off-by: atheik <atteh.mailbox@gmail.com>