ksmbd-tools: install systemd unit file when applicable
The unit file was not installed in systemdsystemunitdir, leaving it up
to the user. Modify the autotools and meson builds such that it is
installed automatically when systemd is present. Generate the unit file
from a template during the build so that paths to the utilities end up
correct. Also, remove scripts/install.sh as it was outdated and now
becomes obsolete.

Signed-off-by: atheik <atteh.mailbox@gmail.com>