ksmbd-tools: fix process exit codes
The exit code macros EXIT_SUCCESS (0) and EXIT_FAILURE (1) are only
meaningful when retuning from main() or when calling exit(). Stop
returning these macros in functions other than main(). Exit codes are
from 0 to 255, inclusive, with non-zero denoting failure. As such,
negated errno is unusual to use as a process exit code so instead use
the generic EXIT_FAILURE macro. This codebase does not propagate errno
from the function that sets it to the return value of the caller,
instead opting for manually choosing an appropriate negated errno to
return. Change a few of them to more relevant ones. While we're at it,
use the errno derived message instead of errno itself in error messages
of ksmbd.control.

Signed-off-by: atheik <atteh.mailbox@gmail.com>