ksmbd-tools: fix spec file
The RPM spec file was adopted from downstream as a template and never
could build with upstream tarballs. It used `%config' macro so as to
protect against changes to `smb.conf.example', even though it's not
something that holds user-made changes. It needlessly required that the
build machine had at least kernel 5.15 even though that should only be
required at package installation time. Fix the above and modify it such
that redundant install target steps are removed, the man pages are
installed, and that sufficient systemd version is checked to be present
at package installation time so as to install a functioning unit file.

Signed-off-by: atheik <atteh.mailbox@gmail.com>