ksmbd-tools: don't treat global guest as share guest on config reload
When the global guest account is specified in ksmbd.conf and a share
guest account is not, the global guest account definition is mistaken
for the share guest account definition on config reload. This happens
due to the `guest account' global parameter and the `guest account'
share parameter having the same name, together with the implemented way
share parameter definitions in the global section become the implicit
defaults for all shares. Fix this by removing `guest account' from the
parameters that are to be applied to all shares. Note that on initial
startup, i.e. not on config reload, it is done in global_conf_create().

On both initial startup and config reload, create the global guest
account in global_conf_fixup_missing(). This fixes another bug, where
the global guest account is not recreated on config reload if both the
global guest account and the share guest account are user-given, caused
partially also by the first bug identified and fixed here.

Finally, add usm_add_guest_account() and use it when needing to add a
guest account, whether it be the global one or the share one. Note that
it is purposefully identical to usm_add_new_user() when it comes to the
use and lifetime of `name'. Also, tweak the debug log message when user
already exists in usm_add_new_user() so that it is less ambiguous when
seen with the following kill_ksmbd_user() debug log message.

Signed-off-by: Atte Heikkilä <atteh.mailbox@gmail.com>