ksmbd-tools: fix user initial UID and GID
In new_ksmbd_user(), the new user has UID and GID of 9999 if the name
does not match a system user. This 9999 has no special meaning and may
result in the user mapping into a system user unintentionally. This is
not the case with 65535 which is historically considered the invalid
UID and GID. In ksmbd-tools and ksmbd this value is hidden behind the
the KSMBD_SHARE_INVALID_UID and KSMBD_SHARE_INVALID_GID macros, which
misleadingly suggest that they are arbitrary values chosen to be
special only in the context of ksmbd.

Signed-off-by: Atte Heikkilä <atteh.mailbox@gmail.com>