ksmbd-tools: introduce add_users_map() and stub add_hosts_map()
Remove parse_list() by introducing add_users_map() and add_hosts_map().
The latter is a stub, which is still a improvement over parse_list()
since it did not support hosts lists at all. Make the functions
responsible for freeing the lists, as was the case previously with
parse_list(), since that is more suitable for add_hosts_map().

Signed-off-by: Atte Heikkilä <atteh.mailbox@gmail.com>