ksmbd-tools: replace --options with --option in addshare
In addshare when using the `--options' option, two share parameters are
delimited by a *single* character. This delimiting character precedes a
share parameter name and is changed into a newline. Extra characters
preceding the delimiter, such as whitespace, become part of the value
of the preceding share parameter and may make it invalid. Also, an
invalid parameter malforms a preceding share parameter as no delimiter
change happens. This type of option is unusual since it delegates the
option splitting to addshare, instead of having the user specify it
explicitly at the shell or whatever calls execve(2). As such, replace
`--options' with `--option', which can be given multiple times.

Signed-off-by: Atte Heikkilä <atteh.mailbox@gmail.com>