ksmbd-tools: simplify addshare and do writes in order
Write shares as well as their parameters in alphabetical order, instead
of the nondescript hash table iteration order, and visually separate
share parameters from all others. If special groups exist, they will
be written first. Build the config file contents into a GPtrArray and
do the write with set_conf_contents() calling g_file_set_contents().
The function makes guarantees for file consistency and in the future
may be used for writing the user database and the subauth file as well.
Get rid of the AUX group and instead modify the group directly when
updating a share. To make this possible, remove the existing key in
cp_parse_external_smbconf_group() before adding the new one with
add_group_key_value(). Also, this allows for unsetting a share
parameter with `--option', which was not possible with `--options'.
Don't leak final allocations in addshare_main() just for good measure.

Signed-off-by: Atte Heikkilä <atteh.mailbox@gmail.com>