ksmbd-tools: simplify adduser and do writes in order
As with addshare, write users in alphabetical order, instead of the
nondescript hash table iteration order. Likewise, build the config file
contents into a GPtrArray and do the write with set_conf_contents().
Also, make the password processing more straightforward by having the
functions in the processing pipeline follow the same prototype and act
on the same arguments. As for password prompting, instead of processing
password input line-by-line in canonical mode, do it character-by-
character in noncanonical mode, so as to bypass the TTY line discipline
since it is rarely useful for password input. ^D is now an error. Don't
leak final allocations in adduser_main() just for good measure.

Signed-off-by: Atte Heikkilä <atteh.mailbox@gmail.com>