ksmbd-tools: add share parameter prompting to addshare
When `--option' is not given for addshare, prompt for share parameters.
The editor works in noncanonical mode and supports erasing characters
as well as cycling through share parameters with the up and down arrow
keys, or alternatively, with C-p and C-n as in Emacs. ANSI-compatible
terminal is required. The enter key may also be used, and when the user
has defined values for the share parameters as desired, going past the
last parameter will commit changes. C-w fully erases the current value.
When updating an existing share, share parameters have their existing
values prefilled for the user to edit. When a share parameter does not
have an existing value, its prefilled value is a comment describing the
parameter with the implicit default value in brackets (`[]'). Tab key
can be used for cycling through available completions. To that end,
parse the user database also in addshare. Also, process groups fully
in addshare since we require the `root directory' global parameter for
path completions.

Signed-off-by: Atte Heikkilä <atteh.mailbox@gmail.com>