ksmbd-tools: make share/user non-options and unify option naming
The addshare and adduser utilities *always* operate on a single share
and a single user, respectively, and so it does not make sense that the
share or the user should be tied to any particular option. As such,
make both of them non-option arguments. As a result, we can now guess
whether to add or update the share or the user if no option is given,
based on whether the share or the user exists. Finally, unify the
naming convention for the options used for overriding the ksmbd.conf
and ksmbdpwd.db file locations. In all of the utilities, the use and
meaning of these options is exactly the same, and so the option names
should not differ.

Signed-off-by: Atte Heikkilä <atteh.mailbox@gmail.com>