ksmbd-tools: call process_entry() once for empty files
process_entry() has to be called once for empty files in case the user
edits the subauth file and leaves it empty. Although an empty subauth
file is an error, an empty pwddb file is *not*. As such, take this into
account by introducing cp_pwddb_eol(). Now, the pwddb file may include
empty lines.

Signed-off-by: Atte Heikkilä <atteh.mailbox@gmail.com>