ksmbd-tools: improve lock file handling
Improve lock file validation and move file parsing to the config parser
altogether. Drop the use of advisory locking since it is in practice
sufficient to just check the PID by sending it the null signal. Also,
don't remove the lock file on exit as that is racy and the job of the
service manager.

Signed-off-by: Atte Heikkilä <atteh.mailbox@gmail.com>