ksmbd-tools: add --list to control
Add `--list' option to control. When the manager receives SIGUSR1, it
passes it to the worker, and then opens the named pipe created by the
sender process. The manager creates a pipe before forking the worker,
with the worker then using the write end to send the config listing to
the manager when signalled to. As for marking the end of a listing, the
worker can't close the write end, as it would prevent further listings,
but will instead send SIGUSR1 back to the manager. Then, the manager
closes the write end of the named pipe, having written to it the
listing received from the worker.

Note that unlike in addshare and adduser, the share and user tables see
concurrent use in mountd and as such we cannot do an ordered write of
the listing without allocs. For now, the shares are listed in the
nondescript hash table iteration order.

Signed-off-by: Atte Heikkilä <atteh.mailbox@gmail.com>