ksmbd-tools: make debug prints config reload friendly
After a config reload, a stale share is killed only after the last
client connected to it has disconnected. Thus, it is impossible to know
which instance of the share a `Kill share ...' message refers to when
viewing the logs. Fix this by printing, along with the share name, a
representation of the `struct ksmbd_share *' pointer. Do this for users
as well.

Signed-off-by: Atte Heikkilä <atteh.mailbox@gmail.com>